Stripe: Continue unwinding the UtilityHookup complexity
