--- app/utilities/null_utility.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class NullUtility < Utility
-end
