--- app/utilities/stripe/stripe_utility.rb
@@ -1,15 +1,0 @@
-module Stripe
-  class StripeUtility < Utility
-    def api_token=api_token
-      configuration["api_token"] = api_token
-    end
-
-    def api_token
-      configuration["api_token"]
-    end
-
-    def attribute_names
-      super + [:api_token]
-    end
-  end
-end
