--- app/utilities/stripe/stripe_utility_policy.rb
@@ -1,7 +1,0 @@
-module Stripe
-  class StripeUtilityPolicy < UtilityPolicy
-    def permitted_attributes(_params)
-      [:name, :api_token]
-    end
-  end
-end
