+++ app/utilities/stripe_utility/_form.html.erb
@@ -1,5 +1,1 @@
-
-<%= form.fields_for(:utility) do |utility_fields| %>
-  <%= render "password_field", form: utility_fields, attribute: :api_token %>
-<%- end %>
-
+<%= render "password_field", form: form, attribute: :api_token %>
