+++ app/utilities/stripe_utility_policy.rb
@@ -1,0 +1,5 @@
+class StripeUtilityPolicy < UtilityHookupPolicy
+  def permitted_attributes(_params)
+    [:name, :api_token]
+  end
+end
