+++ app/views/utility_hookups/_form.html.erb
@@ -1,4 +1,5 @@
 <%= form_with(model: [space, utility_hookup]) do |form| %>
+  <%- form.object = form.object.utility %>
   <%= form.hidden_field :utility_slug %>
   <header>
     <h4>
